Pull delegate plumbing out of turn_coordinator to stay inside the budget gate
The announce-flow work was functionally correct after the sqlite fixes, but it
left turn_coordinator beyond the structural budget enforced by governance. The
extra async delegate helpers were implementation details, not core turn-routing
logic, so keeping them inline made the hottest orchestration file harder to
maintain and blocked CI.

This extracts delegate-support helpers into a dedicated conversation module and
refreshes the tracked architecture drift report so the evidence matches the new
layout. The runtime behavior stays the same; the code just lives at a better
boundary.

Constraint: turn_coordinator must remain within the tracked structural budget
Constraint: delegate terminal persistence and announce behavior must remain unchanged during extraction
Rejected: raise the budget ceiling for turn_coordinator | normalizes continued growth instead of reducing orchestration sprawl
Rejected: leave the helper block inline and rely only on drift-report updates | governance would still fail and maintainability would worsen
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: keep delegate lifecycle helpers in the dedicated support module unless the orchestration control flow itself truly needs inline ownership
Tested: cargo fmt --all -- --check; LOONGCLAW_ARCH_STRICT=true ./scripts/check_architecture_boundaries.sh; ./scripts/generate_architecture_drift_report.sh docs/releases/architecture-drift-2026-04.md; bash scripts/check_architecture_drift_freshness.sh docs/releases/architecture-drift-2026-04.md
Not-tested: local cargo compile/test rerun blocked by concurrent local builds; CI rerun required after push
