+++ crates/app/src/conversation/turn_coordinator.rs
@@ -1,19 +1,9 @@
-#[cfg(feature = "memory-sqlite")]
-use std::any::Any;
 use std::collections::{BTreeMap, BTreeSet};
-#[cfg(feature = "memory-sqlite")]
-use std::future::Future;
-#[cfg(feature = "memory-sqlite")]
-use std::panic::AssertUnwindSafe;
 
 use async_trait::async_trait;
-#[cfg(feature = "memory-sqlite")]
-use futures_util::FutureExt;
 use loongclaw_contracts::{AuditEventKind, ExecutionPlane, PlaneTier};
 use serde::Serialize;
 use serde_json::{Value, json};
-#[cfg(feature = "memory-sqlite")]
-use tokio::runtime::Handle;
 use tokio::sync::Mutex;
 #[cfg(feature = "memory-sqlite")]
 use tokio::time::{Duration, Instant, timeout};
@@ -29,11 +19,9 @@ use crate::acp::{
 use crate::memory::runtime_config::MemoryRuntimeConfig;
 #[cfg(feature = "memory-sqlite")]
 use crate::operator::delegate_runtime::{
-    DelegateChildExecutionPolicy, build_delegate_child_lifecycle_seed, next_delegate_child_depth,
+    DelegateChildExecutionPolicy, build_delegate_child_lifecycle_seed,
 };
 use crate::runtime_self_continuity;
-#[cfg(feature = "memory-sqlite")]
-use crate::session::frozen_result::capture_frozen_result;
 
 use super::super::config::{LoongClawConfig, ToolConsentMode};
 use super::ConversationSessionAddress;
@@ -43,10 +31,18 @@ use super::analytics::{
     TurnCheckpointRecoveryAction, build_turn_checkpoint_repair_plan, summarize_safe_lane_history,
 };
 #[cfg(feature = "memory-sqlite")]
-use super::announce::{DelegateAnnounceSettings, enqueue_delegate_result_announce};
+use super::announce::DelegateAnnounceSettings;
 #[cfg(feature = "memory-sqlite")]
 use super::approval_resolution::CoordinatorApprovalResolutionRuntime;
 use super::context_engine::{AssembledConversationContext, ConversationContextEngine};
+#[cfg(feature = "memory-sqlite")]
+use super::delegate_support::{
+    finalize_and_announce_delegate_child_terminal, finalize_async_delegate_spawn_failure,
+    finalize_async_delegate_spawn_failure_with_recovery,
+    finalize_delegate_child_terminal_with_recovery, format_delegate_child_panic,
+    next_delegate_child_depth_for_delegate, spawn_async_delegate_detached,
+    with_prepared_subagent_spawn_cleanup_if_kernel_bound,
+};
 use super::ingress::ConversationIngressContext;
 use super::lane_arbiter::{ExecutionLane, LaneArbiterPolicy, LaneDecision};
 use super::persistence::{
@@ -3427,7 +3423,7 @@ async fn emit_async_delegate_child_queued_event<R: ConversationRuntime + ?Sized>
 }
 
 #[cfg(feature = "memory-sqlite")]
-async fn emit_async_delegate_child_terminal_event<R: ConversationRuntime + ?Sized>(
+pub(crate) async fn emit_async_delegate_child_terminal_event<R: ConversationRuntime + ?Sized>(
     runtime: &R,
     parent_session_id: &str,
     child_session_id: &str,
@@ -4820,321 +4816,6 @@ pub(crate) async fn run_started_delegate_child_turn_with_runtime<
             }
             Ok(outcome)
         }
-    }
-}
-
-#[cfg(all(feature = "memory-sqlite", test))]
-fn finalize_async_delegate_spawn_failure(
-    memory_config: &MemoryRuntimeConfig,
-    child_session_id: &str,
-    parent_session_id: &str,
-    label: Option<String>,
-    profile: Option<crate::conversation::DelegateBuiltinProfile>,
-    execution: &ConstrainedSubagentExecution,
-    max_frozen_bytes: usize,
-    error: String,
-) -> Result<(), String> {
-    crate::operator::delegate_runtime::finalize_async_delegate_spawn_failure(
-        memory_config,
-        child_session_id,
-        parent_session_id,
-        label,
-        profile,
-        execution,
-        max_frozen_bytes,
-        error,
-    )
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn finalize_async_delegate_spawn_failure_with_recovery(
-    memory_config: &MemoryRuntimeConfig,
-    child_session_id: &str,
-    parent_session_id: &str,
-    label: Option<String>,
-    profile: Option<crate::conversation::DelegateBuiltinProfile>,
-    execution: &ConstrainedSubagentExecution,
-    max_frozen_bytes: usize,
-    error: String,
-) -> Result<(), String> {
-    crate::operator::delegate_runtime::finalize_async_delegate_spawn_failure_with_recovery(
-        memory_config,
-        child_session_id,
-        parent_session_id,
-        label,
-        profile,
-        execution,
-        max_frozen_bytes,
-        error,
-    )
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn format_async_delegate_spawn_panic(panic_payload: Box<dyn Any + Send>) -> String {
-    let panic_payload = match panic_payload.downcast::<String>() {
-        Ok(message) => return format!("delegate_async_spawn_panic: {}", *message),
-        Err(panic_payload) => panic_payload,
-    };
-    match panic_payload.downcast::<&'static str>() {
-        Ok(message) => format!("delegate_async_spawn_panic: {}", *message),
-        Err(_) => "delegate_async_spawn_panic".to_owned(),
-    }
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn spawn_async_delegate_detached(
-    runtime_handle: Handle,
-    config: std::sync::Arc<LoongClawConfig>,
-    memory_config: MemoryRuntimeConfig,
-    spawner: std::sync::Arc<dyn AsyncDelegateSpawner>,
-    request: AsyncDelegateSpawnRequest,
-    max_frozen_bytes: usize,
-    announce_settings: DelegateAnnounceSettings,
-) {
-    let child_session_id = request.child_session_id.clone();
-    let parent_session_id = request.parent_session_id.clone();
-    let label = request.label.clone();
-    let profile = request.profile;
-    let execution = request.execution.clone();
-    let binding = request.binding.clone();
-    runtime_handle.spawn(async move {
-        let spawn_failure = match AssertUnwindSafe(spawner.spawn(request))
-            .catch_unwind()
-            .await
-        {
-            Ok(Ok(())) => None,
-            Ok(Err(error)) => Some(error),
-            Err(panic_payload) => Some(format_async_delegate_spawn_panic(panic_payload)),
-        };
-        if let Some(error) = spawn_failure {
-            let spawn_error = error.clone();
-            let finalize_result = finalize_async_delegate_spawn_failure_with_recovery(
-                &memory_config,
-                &child_session_id,
-                &parent_session_id,
-                label.clone(),
-                profile,
-                &execution,
-                max_frozen_bytes,
-                error.clone(),
-            );
-            if let Err(finalize_error) = &finalize_result {
-                tracing::warn!(
-                    child_session_id = %child_session_id,
-                    parent_session_id = %parent_session_id,
-                    spawn_error = %spawn_error,
-                    error = %finalize_error,
-                    "delegate async spawn failure finalize fell back with error"
-                );
-            }
-            enqueue_delegate_result_announce_with_memory_config(
-                memory_config.clone(),
-                parent_session_id.clone(),
-                child_session_id.clone(),
-                announce_settings.clone(),
-            );
-
-            let runtime = DefaultConversationRuntime::from_config_or_env(config.as_ref());
-            match runtime {
-                Ok(runtime) => {
-                    emit_async_delegate_child_terminal_event(
-                        &runtime,
-                        &parent_session_id,
-                        &child_session_id,
-                        label.as_deref(),
-                        profile,
-                        "failed",
-                        execution.isolation,
-                        0,
-                        None,
-                        Some(error.as_str()),
-                        None,
-                        execution.workspace_root.as_deref(),
-                        None,
-                        binding.as_borrowed(),
-                    )
-                    .await;
-                }
-                Err(runtime_error) => {
-                    tracing::warn!(
-                        child_session_id = %child_session_id,
-                        parent_session_id = %parent_session_id,
-                        error = %runtime_error,
-                        "delegate async spawn failure could not emit parent terminal projection"
-                    );
-                }
-            }
-        }
-    });
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn enqueue_delegate_result_announce_for_parent(
-    config: &LoongClawConfig,
-    parent_session_id: &str,
-    child_session_id: &str,
-) {
-    let memory_config = MemoryRuntimeConfig::from_memory_config(&config.memory);
-    let announce_settings = DelegateAnnounceSettings::from_config(config);
-    enqueue_delegate_result_announce_with_memory_config(
-        memory_config,
-        parent_session_id.to_owned(),
-        child_session_id.to_owned(),
-        announce_settings,
-    );
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn enqueue_delegate_result_announce_with_memory_config(
-    memory_config: MemoryRuntimeConfig,
-    parent_session_id: String,
-    child_session_id: String,
-    announce_settings: DelegateAnnounceSettings,
-) {
-    enqueue_delegate_result_announce(
-        memory_config,
-        parent_session_id,
-        child_session_id,
-        announce_settings,
-    );
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn next_delegate_child_depth_for_delegate(
-    config: &LoongClawConfig,
-    repo: &SessionRepository,
-    session_context: &SessionContext,
-) -> Result<usize, String> {
-    next_delegate_child_depth(
-        repo,
-        &session_context.session_id,
-        config.tools.delegate.max_depth,
-    )
-}
-
-#[cfg(feature = "memory-sqlite")]
-pub(crate) async fn with_prepared_subagent_spawn_cleanup_if_kernel_bound<
-    R: ConversationRuntime + ?Sized,
-    F,
-    Fut,
-    T,
->(
-    runtime: &R,
-    parent_session_id: &str,
-    child_session_id: &str,
-    binding: ConversationRuntimeBinding<'_>,
-    work: F,
-) -> Result<T, String>
-where
-    F: FnOnce() -> Fut,
-    Fut: Future<Output = Result<T, String>>,
-{
-    prepare_subagent_spawn_if_kernel_bound(runtime, parent_session_id, child_session_id, binding)
-        .await?;
-    let work_result = work().await;
-    let notify_result = notify_subagent_ended_if_kernel_bound(
-        runtime,
-        parent_session_id,
-        child_session_id,
-        binding,
-    )
-    .await;
-    match (work_result, notify_result) {
-        (Ok(value), Ok(())) => Ok(value),
-        (Err(work_error), Ok(())) => Err(work_error),
-        (Ok(_), Err(notify_error)) => {
-            Err(format!("delegate_subagent_end_hook_failed: {notify_error}"))
-        }
-        (Err(work_error), Err(notify_error)) => Err(format!(
-            "{work_error}; delegate_subagent_end_hook_failed: {notify_error}"
-        )),
-    }
-}
-
-#[cfg(feature = "memory-sqlite")]
-async fn prepare_subagent_spawn_if_kernel_bound<R: ConversationRuntime + ?Sized>(
-    runtime: &R,
-    parent_session_id: &str,
-    child_session_id: &str,
-    binding: ConversationRuntimeBinding<'_>,
-) -> Result<(), String> {
-    let Some(kernel_ctx) = binding.kernel_context() else {
-        return Ok(());
-    };
-    runtime
-        .prepare_subagent_spawn(parent_session_id, child_session_id, kernel_ctx)
-        .await
-}
-
-#[cfg(feature = "memory-sqlite")]
-async fn notify_subagent_ended_if_kernel_bound<R: ConversationRuntime + ?Sized>(
-    runtime: &R,
-    parent_session_id: &str,
-    child_session_id: &str,
-    binding: ConversationRuntimeBinding<'_>,
-) -> Result<(), String> {
-    let Some(kernel_ctx) = binding.kernel_context() else {
-        return Ok(());
-    };
-    runtime
-        .on_subagent_ended(parent_session_id, child_session_id, kernel_ctx)
-        .await
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn finalize_delegate_child_terminal_with_recovery(
-    repo: &SessionRepository,
-    child_session_id: &str,
-    request: FinalizeSessionTerminalRequest,
-) -> Result<(), String> {
-    crate::operator::delegate_runtime::finalize_delegate_child_terminal_with_recovery(
-        repo,
-        child_session_id,
-        request,
-    )
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn finalize_and_announce_delegate_child_terminal(
-    config: &LoongClawConfig,
-    repo: &SessionRepository,
-    child_session_id: &str,
-    parent_session_id: &str,
-    outcome: &loongclaw_contracts::ToolCoreOutcome,
-    state: SessionState,
-    last_error: Option<String>,
-    event_kind: &str,
-    event_payload_json: Value,
-) -> Result<(), String> {
-    let max_frozen_bytes = config.tools.delegate.max_frozen_bytes;
-    let frozen_result = capture_frozen_result(outcome, max_frozen_bytes);
-
-    let request = FinalizeSessionTerminalRequest {
-        state,
-        last_error,
-        event_kind: event_kind.to_owned(),
-        actor_session_id: Some(parent_session_id.to_owned()),
-        event_payload_json,
-        outcome_status: outcome.status.clone(),
-        outcome_payload_json: outcome.payload.clone(),
-        frozen_result: Some(frozen_result),
-    };
-
-    finalize_delegate_child_terminal_with_recovery(repo, child_session_id, request)?;
-    enqueue_delegate_result_announce_for_parent(config, parent_session_id, child_session_id);
-
-    Ok(())
-}
-
-#[cfg(feature = "memory-sqlite")]
-fn format_delegate_child_panic(panic_payload: Box<dyn Any + Send>) -> String {
-    let panic_payload = match panic_payload.downcast::<String>() {
-        Ok(message) => return format!("delegate_child_panic: {}", *message),
-        Err(panic_payload) => panic_payload,
-    };
-    match panic_payload.downcast::<&'static str>() {
-        Ok(message) => format!("delegate_child_panic: {}", *message),
-        Err(_) => "delegate_child_panic".to_owned(),
     }
 }
 
@@ -7072,6 +6753,10 @@ mod tests {
     use super::*;
     use crate::config::ToolConfig;
     use crate::context::bootstrap_test_kernel_context;
+    use crate::conversation::delegate_support::{
+        finalize_async_delegate_spawn_failure, finalize_async_delegate_spawn_failure_with_recovery,
+        finalize_delegate_child_terminal_with_recovery,
+    };
     use crate::conversation::turn_engine::ToolBatchExecutionIntentTrace;
     use crate::conversation::{
         ConversationTurnObserver, ConversationTurnPhase, ConversationTurnToolState,
